ublox: use only inmX interface name to manage routes
The previous logic, where we would run AT+CGPADDR=[CID] to get the
current IP address assigned to a given CID, and then trying to match
that IP address with one of the gateways in the AT+UIPROUTE? output
isn't good enough due to what seems to be several race conditions:

  * Sometimes AT+CGPADDR=[CID] returns 0.0.0.0.

  * Sometimes AT+CGPADDR=[CID] returns an IP address that doesn't
    match any gateway in the routes reported by AT+UIPROUTE?.

So, instead of trying to match IP addresses, just assume that the inmX
interface is always associated to the same CID, where X=CID-1. This is
true since 40.19 firmware version.