ublox: always add our own custom default route when needed
Newer versions of the TOBY-L4 firmware will only setup a default route
over the first PDP context connected (e.g. the default LTE bearer).

For any other PDP context, we need to be the ones managing all the
routes, including the default route if it needs to have one.

In order to cope with this situation, we will always delete the
default route provided by the module, and add our own afterwards.